# Sprint-1/xpharmacy-be/server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: x-tutor
  description: My cool app
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /products/{id}:
    get:
      tags: 
        - Products
      description: Get a product's info
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: id of product
      responses:
        200: 
          description: Return Product's info
    delete:
      tags:
        - Products
      description: Delete a product
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        200:
          description: Delete success
        404:
          description: Product not found
      
  # Cart
  #*******************************
...
  /orders/{id}:
    get:
      tags: 
        - Orders
      description: Get an order
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: id of user
      responses:
        200: 
          description: Return an Order
    delete:
      tags:
        - Orders
      description: Delete an order
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: id of an order
      responses:
        200:
          description: Delete success
        404:
          description: Order not found
...